7/17 functional

+++ Actor.java
@@ -43,11 +43,11 @@ public abstract class Actor {
         return false;
     }
 
-    public double getX()
+    public int getX()
     {
         return x;
     }
-    public double getY()
+    public int getY()
     {
         return y;
     }
