+++ Shell.java
@@ -2,7 +2,7 @@
 import java.awt.Graphics;
 
 public class Shell extends Actor{
-	public Shell(double x, double y, double angle, int hp)
+	public Shell(int x, int y, double angle, int hp)
 	{
 		super(x, y, angle, hp);
 	}
